# spatialguard/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpatialGuard",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/ring/operations/preferences": {
      "patch": {
        "summary": "Operations Preferences",
        "operationId": "operations_preferences_v1_ring_operations_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationsPreferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/operations/wall": {
      "put": {
        "summary": "Camera Wall",
        "operationId": "camera_wall_v1_ring_operations_wall_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraWall"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/operations/alerts/{alert_id}/acknowledge": {
      "post": {
        "summary": "Acknowledge Alert",
        "operationId": "acknowledge_alert_v1_ring_operations_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/timelapses": {
      "post": {
        "summary": "Create Timelapse",
        "operationId": "create_timelapse_v1_ring_timelapses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelapseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/timelapses/{project_id}/capture": {
      "post": {
        "summary": "Capture Timelapse",
        "operationId": "capture_timelapse_v1_ring_timelapses__project_id__capture_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/timelapses/{project_id}/frames/{frame_id}": {
      "get": {
        "summary": "Timelapse Frame",
        "operationId": "timelapse_frame_v1_ring_timelapses__project_id__frames__frame_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "frame_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Frame Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/timelapses/{project_id}/reel": {
      "get": {
        "summary": "Timelapse Reel",
        "operationId": "timelapse_reel_v1_ring_timelapses__project_id__reel_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/timelapses/{project_id}": {
      "delete": {
        "summary": "Delete Timelapse",
        "operationId": "delete_timelapse_v1_ring_timelapses__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ring/sites/{site}/cameras/{camera}/snapshot": {
      "get": {
        "summary": "Snapshot",
        "operationId": "snapshot_v1_ring_sites__site__cameras__camera__snapshot_get",
        "parameters": [
          {
            "name": "site",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site"
            }
          },
          {
            "name": "camera",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/jpeg": {},
              "image/png": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents/{incident_id}/classify": {
      "post": {
        "summary": "Classify Incident",
        "operationId": "classify_incident_v1_incidents__incident_id__classify_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/devices/{device}/streams": {
      "post": {
        "summary": "Stream",
        "operationId": "stream_v1_ring_devices__device__streams_post",
        "parameters": [
          {
            "name": "device",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamOffer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ring/streams/{sid}": {
      "delete": {
        "summary": "Close",
        "operationId": "close_v1_ring_streams__sid__delete",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Stream State",
        "operationId": "stream_state_v1_ring_streams__sid__get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Association": {
        "properties": {
          "from_observation_id": {
            "type": "string",
            "title": "From Observation Id"
          },
          "to_observation_id": {
            "type": "string",
            "title": "To Observation Id"
          },
          "state": {
            "type": "string",
            "const": "possible",
            "title": "State",
            "default": "possible"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "unobserved_gap_seconds": {
            "type": "number",
            "title": "Unobserved Gap Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "from_observation_id",
          "to_observation_id",
          "reason",
          "unobserved_gap_seconds"
        ],
        "title": "Association"
      },
...
      "CameraWall": {
        "properties": {
          "devices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Devices"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "devices"
        ],
        "title": "CameraWall"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Incident": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "revision_id": {
            "type": "string",
            "title": "Revision Id"
          },
          "calibration_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Calibration Ids",
            "default": []
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "needs_review",
              "reviewed"
            ],
            "title": "Status",
            "default": "needs_review"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "evidence_mode": {
            "type": "string",
            "enum": [
              "replay",
              "live",
              "simulator"
            ],
            "title": "Evidence Mode",
            "default": "replay"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "type": "array",
            "title": "Observations"
          },
          "associations": {
            "items": {
              "$ref": "#/components/schemas/Association"
            },
            "type": "array",
            "title": "Associations"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "classification_status": {
            "type": "string",
            "enum": [
              "not_requested",
              "completed",
              "unavailable"
            ],
            "title": "Classification Status",
            "default": "not_requested"
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentClassification"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "site_id",
          "revision_id",
          "run_id",
          "title",
          "rule",
          "started_at",
          "created_at",
          "observations",
          "associations",
          "evidence_ids"
        ],
        "title": "Incident"
      },
      "IncidentClassification": {
        "properties": {
          "label": {
            "type": "string",
            "enum": [
              "delivery_activity",
              "face_covering_visible",
              "possible_weapon_visible",
              "possible_unauthorized_entry",
              "unidentified_person",
              "package_visible",
              "animal_visible",
              "vehicle_visible",
              "unclear",
              "no_relevant_activity"
            ],
            "title": "Label"
          },
          "display_label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Label"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence"
          },
          "summary": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Summary"
          },
          "visible_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Visible Evidence"
          },
          "uncertainty": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Uncertainty"
          },
          "model": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model"
          },
          "response_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Id"
          },
          "analyzed_at": {
            "type": "string",
            "title": "Analyzed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "display_label",
          "confidence",
          "summary",
          "visible_evidence",
          "uncertainty",
          "model",
          "analyzed_at"
        ],
        "title": "IncidentClassification"
      },
...
      "Observation": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "0.1",
            "title": "Schema Version",
            "default": "0.1"
          },
          "observation_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Observation Id"
          },
          "site_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Site Id"
          },
          "revision_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Revision Id"
          },
          "source_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Source Id"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Category"
          },
          "location": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PointLocation"
              },
              {
                "$ref": "#/components/schemas/RoomLocation"
              },
              {
                "$ref": "#/components/schemas/UnknownLocation"
              }
            ],
            "title": "Location",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "floor_point": "#/components/schemas/PointLocation",
                "room": "#/components/schemas/RoomLocation",
                "unknown": "#/components/schemas/UnknownLocation"
              }
            }
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "observation_id",
          "site_id",
          "revision_id",
          "source_id",
          "observed_at",
          "received_at",
          "category",
          "location",
          "evidence",
          "provenance"
        ],
        "title": "Observation"
      },
      "OperationsPreferences": {
        "properties": {
          "delay_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Delay Seconds",
            "default": 0
          },
          "browser_enabled": {
            "type": "boolean",
            "title": "Browser Enabled",
            "default": false
          },
          "email_enabled": {
            "type": "boolean",
            "title": "Email Enabled",
            "default": false
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "title": "Email",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OperationsPreferences"
      },
...
      "StreamAnswer": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sdp": {
            "type": "string",
            "title": "Sdp"
          },
          "expires_at": {
            "type": "number",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "sdp",
          "expires_at"
        ],
        "title": "StreamAnswer"
      },
      "StreamOffer": {
        "properties": {
          "sdp": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 10,
            "title": "Sdp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sdp"
        ],
        "title": "StreamOffer"
      },
      "StreamState": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "expires_at": {
            "type": "number",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "expires_at"
        ],
        "title": "StreamState"
      },
...
      "TimelapseInput": {
        "properties": {
          "device": {
            "type": "string",
            "title": "Device"
          },
          "site": {
            "type": "string",
            "title": "Site"
          },
          "camera": {
            "type": "string",
            "title": "Camera"
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "cadence_minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 5.0,
            "title": "Cadence Minutes"
          },
          "start_hour": {
            "type": "integer",
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Start Hour"
          },
          "end_hour": {
            "type": "integer",
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "End Hour"
          },
          "timezone": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Timezone",
            "default": "America/Phoenix"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device",
          "site",
          "camera",
          "name",
          "cadence_minutes",
          "start_hour",
          "end_hour"
        ],
        "title": "TimelapseInput"
      },
      "UnknownLocation": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "unknown",
            "title": "Kind"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "reason"
        ],
        "title": "UnknownLocation"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}